+++ project1/src/main/parse/Makefile
@@ -1,6 +1,5 @@
 all: depend
 	gcc checkSyntax.tab.c lex.csy.c checkAccess.tab.c lex.cac.c -c
-	gcc checkSyntax.tab.o lex.csy.o checkAccess.tab.o lex.cac.o testProg.c -o testProg
 
 depend: bison flex
 
@@ -13,4 +12,4 @@ flex:
 	flex -Pcac accessTokens2.l
 
 clean:
-	rm checkSyntax.tab.c checkSyntax.tab.h lex.csy.c checkAccess.tab.c checkAccess.tab.h lex.cac.c checkSyntax.tab.o lex.csy.o checkAccess.tab.o lex.cac.o testProg
+	rm checkSyntax.tab.c checkSyntax.tab.h lex.csy.c checkAccess.tab.c checkAccess.tab.h lex.cac.c checkSyntax.tab.o lex.csy.o checkAccess.tab.o lex.cac.o
