--- project1/src/main/parse/testfile.txt
@@ -1,7 +1,0 @@
-#COMMENT COMMENT
-username1 r
-hello w
-#COMMENT
-username2 w
-username3 b
-#comment
